Extract Attachable::ForNonEditionableModel concern
I'm about to move the content & attachments for a WorldwideOrganisation
from the CorporateInformationPage about page onto the
WorldwideOrganisation itself. Once that move is completed,
WorldwideOrganisation will need to include the Attachable concern and
implement the methods that it requires. PolicyGroup is a non-editionable
model which already supports attachments and it seems to make sense to
re-use its implmentation of the methods that Attachable requires.
Extracting this new concern for non-editionable models will mean the
change to WorldwideOrganisation will be easier.

Note that PolicyGroup did not implement the #organisations method;
whereas Attachable::Null did. Although it looks as if this might only be
used in Reports::PublishedAttachmentsReport which is restricted to
editions with attachments, I decided to implement it on
ForNonEditionableModel for completeness.
